feat(brainstorming): add section on when to go backward
This extracted from work in the original obra/superpowers-skills repo
and modifications made by @jhelwig, credit to both!

This was developed in part with Claude by having it read the modified
skill file, calling out the `When to revisit earlier phases` section,
and having it use the `superpowers:writing-skills` skill to make
suggestions. Finally, the changes were tested with the
`superpowers:testing-skills-with-subagensts` skill which then proposed
further refinements.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>